Constructs tours of data space based on centers of clusters

Constructs tours of data space based on centers of clusters

kmeansPath(data, length = 10, reorder = TRUE, conditionvars = NULL, ...)

pamPath(
  data,
  length = 10,
  reorder = TRUE,
  conditionvars = NULL,
  maxn = 4000,
  ...
)

claraPath(data, length = 10, reorder = TRUE, conditionvars = NULL, ...)

medoidPath(data, cl, reorder = FALSE)

centroidPath(data, cl, reorder = FALSE)

Arguments

data
A dataframe

length
Path length, defaults to 10

reorder
If TRUE uses DendSer to reorder the path dser

conditionvars
A vector of variable names. The returned tour is for this subset of variables.

...
ignored

maxn
(pamPath only) For datasets with more than maxn rows, use maxn randomly selected rows.

cl
A vector specifying cluster membership for rows of data.

Value

A dataframe with the path

Functions

kmeansPath: Constructs a tour of data space following length k-means centroids

pamPath: Constructs a tour of data space following length pam medoids

claraPath: Constructs a tour of data space following length clara medoids

medoidPath: Returns a path visiting cluster medoids

centroidPath: Returns a path visiting cluster centroids

Examples

kmeansPath(mtcars,length=4)

#>        mpg      cyl     disp       hp     drat       wt     qsec           vs
#> 3 15.05000 8.000000 357.6167 194.1667 3.120833 4.104083 17.14250 0.000000e+00
#> 4 18.50000 6.800000 223.4000 198.8000 3.836000 3.001000 15.61600 5.551115e-17
#> 1 20.74286 5.142857 175.1143 102.1429 3.570000 3.194286 19.96714 1.000000e+00
#> 2 28.07500 4.000000  93.6125  81.8750 4.183750 2.042250 18.45000 8.750000e-01
#>              am     gear     carb
#> 3 -1.110223e-16 3.000000 3.083333
#> 4  1.000000e+00 4.600000 5.200000
#> 1  0.000000e+00 3.571429 2.142857
#> 2  1.000000e+00 4.250000 1.500000

pamPath(mtcars,length=4)

#>                      mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> Fiat X1-9           27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1
#> Mazda RX4 Wag       21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> Merc 450SL          17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> Lincoln Continental 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4

claraPath(mtcars,length=4)

#>                      mpg cyl  disp  hp drat    wt  qsec vs am gear carb
#> Fiat X1-9           27.3   4  79.0  66 4.08 1.935 18.90  1  1    4    1
#> Mazda RX4 Wag       21.0   6 160.0 110 3.90 2.875 17.02  0  1    4    4
#> Merc 450SL          17.3   8 275.8 180 3.07 3.730 17.60  0  0    3    3
#> Lincoln Continental 10.4   8 460.0 215 3.00 5.424 17.82  0  0    3    4

medoidPath(mtcars,cl=rep(1:3, length.out=nrow(mtcars)))

#>                mpg cyl  disp  hp drat   wt  qsec vs am gear carb
#> Mazda RX4     17.3   8 275.8 180 3.07 3.73 17.60  0  0    3    3
#> Mazda RX4 Wag 17.8   6 167.6 123 3.92 3.44 18.90  1  0    4    4
#> Datsun 710    22.8   4 108.0  93 3.85 2.32 18.61  1  1    4    1